Add files via upload
First version of a Python program that uses a regular expression for checking if a number is a prime number. It is not a preferred method to check primes. It is just a way to show what regular expressions are capable of and how short and simple this can be done using a regular expression. However, this is just for demo purposes. This way of testing for prime numbers is extremely inefficient and resource demanding. It should never be used in production code.